+++ src/fheroes2/ai/normal/ai_normal_hero.cpp
@@ -50,7 +50,7 @@
 #include "logging.h"
 #include "luck.h"
 #include "maps.h"
-#include "maps_tiles.h"
+#include "maps_tiles_helper.h"
 #include "math_base.h"
 #include "monster.h"
 #include "morale.h"
