+++ src/fheroes2/ai/normal/ai_normal_kingdom.cpp
@@ -47,7 +47,7 @@
 #include "kingdom.h"
 #include "logging.h"
 #include "maps.h"
-#include "maps_tiles.h"
+#include "maps_tiles_helper.h"
 #include "mp2.h"
 #include "mus.h"
 #include "pairs.h"
