+++ src/fheroes2/dialog/dialog_quickinfo.cpp
@@ -50,7 +50,7 @@
 #include "localevent.h"
 #include "logging.h"
 #include "maps.h"
-#include "maps_tiles.h"
+#include "maps_tiles_helper.h"
 #include "math_base.h"
 #include "mp2.h"
 #include "pairs.h"
