+++ src/fheroes2/game/game_startgame.cpp
@@ -64,7 +64,7 @@
 #include "logging.h"
 #include "m82.h"
 #include "maps.h"
-#include "maps_tiles.h"
+#include "maps_tiles_helper.h"
 #include "math_base.h"
 #include "monster.h"
 #include "mp2.h"
