+++ src/fheroes2/kingdom/kingdom.cpp
@@ -46,7 +46,7 @@
 #include "logging.h"
 #include "maps.h"
 #include "maps_fileinfo.h"
-#include "maps_tiles.h"
+#include "maps_tiles_helper.h"
 #include "math_base.h"
 #include "payment.h"
 #include "players.h"
