+++ src/fheroes2/kingdom/view_world.cpp
@@ -41,7 +41,7 @@
 #include "interface_radar.h"
 #include "localevent.h"
 #include "maps.h"
-#include "maps_tiles.h"
+#include "maps_tiles_helper.h"
 #include "mp2.h"
 #include "pairs.h"
 #include "resource.h"
