+++ src/fheroes2/world/world_pathfinding.cpp
@@ -39,7 +39,7 @@
 #include "heroes.h"
 #include "kingdom.h"
 #include "maps.h"
-#include "maps_tiles.h"
+#include "maps_tiles_helper.h"
 #include "math_base.h"
 #include "pairs.h"
 #include "rand.h"
